Duplicate managed-kafka depenedency for core compatibility
We used to declare managed-kafka in the core catalog.  We want to remove this for ease of
maintenance, but this means we would break compatibility as here it's managed-kafka-clients.

This PR adds a temporary duplication of the catalog to keep managed-kafka in the catalog here

This can be removed for 4.0.0

Removal was here https://github.com/micronaut-projects/micronaut-core/pull/7123