Update kafkaErrors.adoc
Existing doc on error handling strategies states that using `RETRY_ON_ERROR` will retry consuming events indefinitely. That does not seem to square with the source of the `@ErrorStrategy` annotation where the default behaviour is that only a single retry is attempted unless the `retryCount` property is set to some other value.