docs(OffsetStrategy): align with docs and implementation
The documentation for `SYNC_PER_RECORD` and `ASYNC_PER_RECORD` were out of sync with other messaging. This change aligns the wording with the wording used in `SYNC` and `ASYNC` as well as with the official documentation for the micronaut-kafka module using the term "processed" to mean after a KafkaListener handles and acts on a ConsumerRecord.

official documentation: https://micronaut-projects.github.io/micronaut-kafka/4.5.2/guide/#kafkaOffsets