Fix make KStream health result serializable (#1026) (#1088)
## Summary
This pull request addresses issue #1026 by implementing a custom `Serde<TaskId>` to kafka-streams module. This avoid a serialization error when calling the `/health` endpoint, which throws in turn a:
```
No serializable introspection present for type TaskId. Consider adding Serdeable. Serializable annotate to type TaskId. Alternatively if you are not in control of the project's source code, you can use @SerdeImport(TaskId.class) to enable serialization of this type.
```

## Changes Made
- Added serde processor as `annotationProcessor`, and serde api as `compileOnly` to load in the annotation. 
- Created a class dedicated to adding annotation for external classes that need serialising as part of the health result

## Related Issues
- Fixes #1026